django/requirements.txt
Cython==0.27.3

Django==1.11.6
django-allauth==0.32.0
django-rest-auth==0.9.2
django-lazysignup==1.1.2
django-appconf==1.0.2
django-compressor==2.2
djangorestframework==3.6.4
django-storages==1.6.5
django-widget-tweaks==1.3
pytest-django==3.1.2
mysqlclient==1.3.12
SQLAlchemy==1.1.15

oauthlib==0.6.1
python-openid==2.2.5
requests==2.18.4
requests-oauthlib==0.4.0

ipython==2.2.0
six==1.11.0
argparse==1.2.1
bidict==0.1.5
filemagic

beautifulsoup4==4.3.2
lxml==3.4.4

dnspython==1.15.0
gsutil==4.28
google-cloud==0.29.0

# Elasticsearch 5.x
elasticsearch>=5.0.0,<6.0.0

# Sphinx
sphinx
# Sphinx "Read the Docs" theme
sphinx_rtd_theme

git+ssh://git@github.com/syscall7/python-bfd.git

git+ssh://git@github.com/williballenthin/python-idb.git@b8223c2c524683359c5d7425bfed28abf15987d8#egg=python-idb
